requirements.txt
ccxt
yfinance
pandas
numpy
requests
google-genai